서버 구동 시 boardStatistic 정보를 맞게 업데이트 해주는 로직 추가
